backend/requirements.txt

Django>=4.2
psycopg2-binary>=2.9
djangorestframework>=3.14
requests>=2.31
python-docx>=1.1
PyPDF2>=3.0
openai
python-dotenv
opencc-python-reimplemented>=0.1.7
python-dateutil>=2.8.2
